# internal/dto/types.gen.go
// Package dto provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.16.3 DO NOT EDIT.
package dto

import (
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for TaskResponseStatus.
const (
	Created    TaskResponseStatus = "created"
	Done       TaskResponseStatus = "done"
	Failed     TaskResponseStatus = "failed"
	Processing TaskResponseStatus = "processing"
)

type CreateTaskRequest struct {
	Title       *string `json:"title,omitempty"`
	Description *string `json:"description,omitempty"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Code    *int    `json:"code,omitempty"`
	Message *string `json:"message,omitempty"`
}

// SuccessResponse defines model for SuccessResponse.
type SuccessResponse struct {
	Code    *int    `json:"code,omitempty"`
	Message *string `json:"message,omitempty"`
}

// TaskResponse defines model for TaskResponse.
type TaskResponse struct {
	Description *string             `json:"description,omitempty"`
	Id          *openapi_types.UUID `json:"id,omitempty"`
	Status      *TaskResponseStatus `json:"status,omitempty"`
	Title       *string             `json:"title,omitempty"`
}

// TaskResponseStatus defines model for TaskResponse.Status.
type TaskResponseStatus string

// GetTasksParams defines parameters for GetTasks.
type GetTasksParams struct {
	// Page Optional. Must be used together with 'amount'.
	Page *int `form:"page,omitempty" json:"page,omitempty"`

	// Amount Optional. Must be used together with 'page'.
	Amount *int `form:"amount,omitempty" json:"amount,omitempty"`
}

// PostTasksJSONBody defines parameters for PostTasks.
type PostTasksJSONBody = CreateTaskRequest

// PatchTasksIdStatusParams defines parameters for PatchTasksIdStatus.
type PatchTasksIdStatusParams struct {
	Status string `form:"status" json:"status"`
}

// PostTasksJSONRequestBody defines body for PostTasks for application/json ContentType.
type PostTasksJSONRequestBody = PostTasksJSONBody